updating readme file, reorganizing file locations
