+++ performance-hints/critical-images/wp-rocket-disable-critical-images-optimization/README.md
