+++ performance-hints/wp-rocket-perf-hints-parameters/README.md
@@ -1,26 +1,34 @@
 # WP Rocket | Performance Hints Parameters
+<br>
 
-Changes the following Performance Hints Optimization parameters:
+### Changes the following Performance Hints Optimization parameters:
 - Warmup links number
 - Width and height thresholds for both desktop and mobile
 - Delay before the beacon script is triggered
 
-Changes the following Optimize Critical Images parameters
+<br>
+
+### Changes the following Optimize Critical Images parameters
 - Elements considered for the optimization
 - Cleanup interval
 
-📝&#160;&#160;**Manual code edit required before use!**
+<br>
 
-Change: 
+### 📝Manual code edit required before use!
 - Only the lines within PROCKETHELPERS_PHO_PARAMETERS. There are notes (// EDIT HERE) showing where to make changes.
 
+<br>
 
-Documentation:
+### Documentation:
 * [Optimize critical images](https://docs.wp-rocket.me/article/1816-optimize-critical-images)
 
-To be used with:
+<br>
+
+### To be used with:
 * Any setup where the default critical images settings do not work optimally.
 
-Last tested with:
+<br>
+
+### Last tested with:
 * WP Rocket 3.16.4
 * WordPress 6.6.x
